Address review comments for #205
Resolves:
- https://github.com/grafana/xk6-browser/pull/205#discussion_r796333349
- https://github.com/grafana/xk6-browser/pull/205#discussion_r796343355
- https://github.com/grafana/xk6-browser/pull/205#discussion_r796335137
- https://github.com/grafana/xk6-browser/pull/205#discussion_r796333615
- https://github.com/grafana/xk6-browser/pull/205#discussion_r796325951
- https://github.com/grafana/xk6-browser/pull/205#discussion_r796291283
- https://github.com/grafana/xk6-browser/pull/205#discussion_r796289881
