+++ common/network_manager.go
@@ -38,6 +38,7 @@ import (
 	k6common "go.k6.io/k6/js/common"
 	k6lib "go.k6.io/k6/lib"
 	k6netext "go.k6.io/k6/lib/netext"
+	k6types "go.k6.io/k6/lib/types"
 	k6stats "go.k6.io/k6/stats"
 )
 
@@ -77,7 +78,7 @@ func NewNetworkManager(
 
 	resolver, err := newResolver(state.Options.DNS)
 	if err != nil {
-		return nil, err
+		return nil, fmt.Errorf("newResolver(%+v): %w", state.Options.DNS, err)
 	}
 
 	m := NetworkManager{
@@ -105,6 +106,50 @@ func NewNetworkManager(
 		return nil, err
 	}
 	return &m, nil
+}
+
+// Returns a new Resolver.
+// Copied with minor changes from
+// https://github.com/grafana/k6/blob/fb70bc6f3d3f22a40e65f32deea3cea1b6d70a76/js/runner.go#L459
+func newResolver(conf k6types.DNSConfig) (k6netext.Resolver, error) {
+	ttl, err := parseTTL(conf.TTL.String)
+	if err != nil {
+		return nil, fmt.Errorf("cannot parse TTL: %w", err)
+	}
+
+	dnsSel := conf.Select
+	if !dnsSel.Valid {
+		dnsSel = k6types.DefaultDNSConfig().Select
+	}
+	dnsPol := conf.Policy
+	if !dnsPol.Valid {
+		dnsPol = k6types.DefaultDNSConfig().Policy
+	}
+	return k6netext.NewResolver(
+		net.LookupIP, ttl, dnsSel.DNSSelect, dnsPol.DNSPolicy), nil
+}
+
+// Parse a string representation of TTL to time.Duration.
+// Copied from https://github.com/grafana/k6/blob/fb70bc6f3d3f22a40e65f32deea3cea1b6d70a76/js/runner.go#L479
+func parseTTL(ttlS string) (time.Duration, error) {
+	ttl := time.Duration(0)
+	switch ttlS {
+	case "inf":
+		// cache "infinitely"
+		ttl = time.Hour * 24 * 365
+	case "0":
+		// disable cache
+	case "":
+		ttlS = k6types.DefaultDNSConfig().TTL.String
+		fallthrough
+	default:
+		var err error
+		ttl, err = k6types.ParseExtendedDuration(ttlS)
+		if ttl < 0 || err != nil {
+			return ttl, fmt.Errorf("invalid DNS TTL: %s", ttlS)
+		}
+	}
+	return ttl, nil
 }
 
 func (m *NetworkManager) deleteRequestByID(reqID network.RequestID) {
@@ -391,61 +436,77 @@ func (m *NetworkManager) onRequestPaused(event *fetch.EventRequestPaused) {
 	defer m.logger.Debugf("NetworkManager:onRequestPaused:return",
 		"sid:%s url:%v", m.session.ID(), event.Request.URL)
 
-	var (
-		failErr error
-		state   = k6lib.GetState(m.ctx)
-	)
+	var failErr error
 
-	defer func() { m.failOrContinueRequest(event, failErr) }()
+	defer func() {
+		if failErr != nil {
+			action := fetch.FailRequest(event.RequestID, network.ErrorReasonBlockedByClient)
+			if err := action.Do(cdp.WithExecutor(m.ctx, m.session)); err != nil {
+				m.logger.Errorf("NetworkManager:onRequestPaused",
+					"error interrupting request: %s", err)
+			} else {
+				m.logger.Warnf("NetworkManager:onRequestPaused",
+					"request %s %s was interrupted: %s", event.Request.Method, event.Request.URL, failErr)
+				return
+			}
+		}
+		action := fetch.ContinueRequest(event.RequestID)
+		if err := action.Do(cdp.WithExecutor(m.ctx, m.session)); err != nil {
+			m.logger.Errorf("NetworkManager:onRequestPaused",
+				"error continuing request: %s", err)
+		}
+	}()
 
 	purl, err := url.Parse(event.Request.URL)
 	if err != nil {
 		m.logger.Errorf("NetworkManager:onRequestPaused",
-			"error parsing URL %q: %s", event.Request.URL, err.Error())
+			"error parsing URL %q: %s", event.Request.URL, err)
 		return
 	}
 
-	host, _, err := net.SplitHostPort(purl.Host)
-	if err != nil {
-		// no port, fallback to original
-		host = purl.Host
-	}
-
-	ip := net.ParseIP(host)
+	var (
+		host  = purl.Hostname()
+		ip    = net.ParseIP(host)
+		state = k6lib.GetState(m.ctx)
+	)
 	if ip != nil {
 		failErr = checkBlockedIPs(ip, state.Options.BlacklistIPs)
 		return
 	}
 	failErr = checkBlockedHosts(host, state.Options.BlockedHostnames.Trie)
-	if failErr == nil {
-		// Do one last check of the resolved IP
-		ip, err = m.resolver.LookupIP(host)
-		if err != nil {
-			m.logger.Debugf("NetworkManager:onRequestPaused",
-				"error resolving %q: %s", host, err.Error())
-			return
-		}
-		failErr = checkBlockedIPs(ip, state.Options.BlacklistIPs)
+	if failErr != nil {
+		return
 	}
+
+	// Do one last check of the resolved IP
+	ip, err = m.resolver.LookupIP(host)
+	if err != nil {
+		m.logger.Debugf("NetworkManager:onRequestPaused",
+			"error resolving %q: %s", host, err)
+		return
+	}
+	failErr = checkBlockedIPs(ip, state.Options.BlacklistIPs)
 }
 
-func (m *NetworkManager) failOrContinueRequest(event *fetch.EventRequestPaused, failErr error) {
-	if failErr != nil {
-		action := fetch.FailRequest(event.RequestID, network.ErrorReasonBlockedByClient)
-		if err := action.Do(cdp.WithExecutor(m.ctx, m.session)); err != nil {
-			m.logger.Errorf("NetworkManager:onRequestPaused",
-				"error interrupting request: %s", err.Error())
-		} else {
-			m.logger.Warnf("NetworkManager:onRequestPaused",
-				"request %s %s was interrupted: %s", event.Request.Method, event.Request.URL, failErr)
-			return
+func checkBlockedHosts(host string, blockedHosts *k6types.HostnameTrie) error {
+	if blockedHosts == nil {
+		return nil
+	}
+	if match, blocked := blockedHosts.Contains(host); blocked {
+		return fmt.Errorf("hostname %s is in a blocked pattern %q", host, match)
+	}
+	return nil
+}
+
+func checkBlockedIPs(ip net.IP, blockedIPs []*k6lib.IPNet) error {
+	for _, ipnet := range blockedIPs {
+		if ipnet.Contains(ip) {
+			// TODO: Return netext.BlackListedIPError here once its private
+			// fields are exported, or there's a constructor for it.
+			return fmt.Errorf("IP %s is in a blacklisted range %q", ip, ipnet)
 		}
 	}
-	action := fetch.ContinueRequest(event.RequestID)
-	if err := action.Do(cdp.WithExecutor(m.ctx, m.session)); err != nil {
-		m.logger.Errorf("NetworkManager:onRequestPaused",
-			"error continuing request: %s", err.Error())
-	}
+	return nil
 }
 
 func (m *NetworkManager) onRequestServedFromCache(event *network.EventRequestServedFromCache) {
